finishes instanceof. Do not declare section .data because it confuses gdb

+++ Compiler/src/cs444/codegen/SelectorIndexedTable.java
@@ -27,7 +27,7 @@ public class SelectorIndexedTable {
         table.addRow(selector);
     }
 
-    public int getOffset(String selector){
+    public long getOffset(String selector){
         return table.getOffset(selector);
     }
